Makefile: use ruby checksum script to ensure build script integrity
This utilities Ruby, which exists on all OS X systems (And most Linuxes)
to download and sha256 check the downloaded script to ensure it matches
the intended checksum, which adds an appreciable level of security
over the status quo.

The script is pretty self-explanatory, but it will comfortable work
with either cURL or wget. The SHA check is an internal Ruby mechanism,
and all the required aspects are natively shipped in every Ruby version
OS X supplies, back to at least 10.6, and potentially earlier.

Should Ruby find the checksum to match the expected checksum, it will
chmod the file with executable permissions and then hand it over to
bash to run. For extra user awareness it will print the checksum
expected and the actual checksum, and then add a "successful verification"
message.

Should Ruby find the checksum failing it immediately erases the file
and puts a warning the checksum failed, and then exits.

This isn't perfect security, it doesn't pretend to be, but it is
a not insigificant step up from the current situation, and may address
some of the concerns expressed by Werner on the GPG Mailing lists today
and hopefully reacts suitably to some of Lukas' comments as well:

https://lists.gnupg.org/pipermail/gnupg-users/2015-February/052574.html
https://lists.gnupg.org/pipermail/gnupg-users/2015-February/052595.html

The plan to remove the remote execution entirely is a good step, and
this can still work with that mechanism very easily, but I thought
right nowrather than watch drama pile up on the GPG Mailing list I'd
try and draw up even a minor improvement to address some of the comments
and hopefully be useful in that way.